pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-agent"
version = "0.1.0"
description = "Telegram bot that recommends what to wear based on weather (Open-Meteo + LangChain)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "deepeval>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit_mock: unit tests with no LLM/HTTP (deterministic)",
    "unit_llm: unit tests with fake LLM (GenericFakeChatModel)",
    "integration_mock: integration with mocked LLM/HTTP",
    "integration_llm: integration with real or eval LLM (needs API key)",
    "system_mock: E2E with fake agent (no real API)",
    "system_llm: E2E with real agent and safety (needs API key)",
    "safety: safety tests (injection, leakage, abuse)",
]

[tool.setuptools.packages.find]
where = ["src"]
